third_party/WebKit/LayoutTests/inspector/console/console-viewport-stick-to-bottom-expected.txt

CONSOLE MESSAGE: line 9: Multiline
Message #0
CONSOLE MESSAGE: line 9: Multiline
Message #1
CONSOLE MESSAGE: line 9: Multiline
Message #2
CONSOLE MESSAGE: line 9: Multiline
Message #3
CONSOLE MESSAGE: line 9: Multiline
Message #4
CONSOLE MESSAGE: line 9: Multiline
Message #5
CONSOLE MESSAGE: line 9: Multiline
Message #6
CONSOLE MESSAGE: line 9: Multiline
Message #7
CONSOLE MESSAGE: line 9: Multiline
Message #8
CONSOLE MESSAGE: line 9: Multiline
Message #9
CONSOLE MESSAGE: line 9: Multiline
Message #10
CONSOLE MESSAGE: line 9: Multiline
Message #11
CONSOLE MESSAGE: line 9: Multiline
Message #12
CONSOLE MESSAGE: line 9: Multiline
Message #13
CONSOLE MESSAGE: line 9: Multiline
Message #14
CONSOLE MESSAGE: line 9: Multiline
Message #15
CONSOLE MESSAGE: line 9: Multiline
Message #16
CONSOLE MESSAGE: line 9: Multiline
Message #17
CONSOLE MESSAGE: line 9: Multiline
Message #18
CONSOLE MESSAGE: line 9: Multiline
Message #19
CONSOLE MESSAGE: line 9: Multiline
Message #20
CONSOLE MESSAGE: line 9: Multiline
Message #21
CONSOLE MESSAGE: line 9: Multiline
Message #22
CONSOLE MESSAGE: line 9: Multiline
Message #23
CONSOLE MESSAGE: line 9: Multiline
Message #24
CONSOLE MESSAGE: line 9: Multiline
Message #25
CONSOLE MESSAGE: line 9: Multiline
Message #26
CONSOLE MESSAGE: line 9: Multiline
Message #27
CONSOLE MESSAGE: line 9: Multiline
Message #28
CONSOLE MESSAGE: line 9: Multiline
Message #29
CONSOLE MESSAGE: line 9: Multiline
Message #30
CONSOLE MESSAGE: line 9: Multiline
Message #31
CONSOLE MESSAGE: line 9: Multiline
Message #32
CONSOLE MESSAGE: line 9: Multiline
Message #33
CONSOLE MESSAGE: line 9: Multiline
Message #34
CONSOLE MESSAGE: line 9: Multiline
Message #35
CONSOLE MESSAGE: line 9: Multiline
Message #36
CONSOLE MESSAGE: line 9: Multiline
Message #37
CONSOLE MESSAGE: line 9: Multiline
Message #38
CONSOLE MESSAGE: line 9: Multiline
Message #39
CONSOLE MESSAGE: line 9: Multiline
Message #40
CONSOLE MESSAGE: line 9: Multiline
Message #41
CONSOLE MESSAGE: line 9: Multiline
Message #42
CONSOLE MESSAGE: line 9: Multiline
Message #43
CONSOLE MESSAGE: line 9: Multiline
Message #44
CONSOLE MESSAGE: line 9: Multiline
Message #45
CONSOLE MESSAGE: line 9: Multiline
Message #46
CONSOLE MESSAGE: line 9: Multiline
Message #47
CONSOLE MESSAGE: line 9: Multiline
Message #48
CONSOLE MESSAGE: line 9: Multiline
Message #49
CONSOLE MESSAGE: line 9: Multiline
Message #50
CONSOLE MESSAGE: line 9: Multiline
Message #51
CONSOLE MESSAGE: line 9: Multiline
Message #52
CONSOLE MESSAGE: line 9: Multiline
Message #53
CONSOLE MESSAGE: line 9: Multiline
Message #54
CONSOLE MESSAGE: line 9: Multiline
Message #55
CONSOLE MESSAGE: line 9: Multiline
Message #56
CONSOLE MESSAGE: line 9: Multiline
Message #57
CONSOLE MESSAGE: line 9: Multiline
Message #58
CONSOLE MESSAGE: line 9: Multiline
Message #59
CONSOLE MESSAGE: line 9: Multiline
Message #60
CONSOLE MESSAGE: line 9: Multiline
Message #61
CONSOLE MESSAGE: line 9: Multiline
Message #62
CONSOLE MESSAGE: line 9: Multiline
Message #63
CONSOLE MESSAGE: line 9: Multiline
Message #64
CONSOLE MESSAGE: line 9: Multiline
Message #65
CONSOLE MESSAGE: line 9: Multiline
Message #66
CONSOLE MESSAGE: line 9: Multiline
Message #67
CONSOLE MESSAGE: line 9: Multiline
Message #68
CONSOLE MESSAGE: line 9: Multiline
Message #69
CONSOLE MESSAGE: line 9: Multiline
Message #70
CONSOLE MESSAGE: line 9: Multiline
Message #71
CONSOLE MESSAGE: line 9: Multiline
Message #72
CONSOLE MESSAGE: line 9: Multiline
Message #73
CONSOLE MESSAGE: line 9: Multiline
Message #74
CONSOLE MESSAGE: line 9: Multiline
Message #75
CONSOLE MESSAGE: line 9: Multiline
Message #76
CONSOLE MESSAGE: line 9: Multiline
Message #77
CONSOLE MESSAGE: line 9: Multiline
Message #78
CONSOLE MESSAGE: line 9: Multiline
Message #79
CONSOLE MESSAGE: line 9: Multiline
Message #80
CONSOLE MESSAGE: line 9: Multiline
Message #81
CONSOLE MESSAGE: line 9: Multiline
Message #82
CONSOLE MESSAGE: line 9: Multiline
Message #83
CONSOLE MESSAGE: line 9: Multiline
Message #84
CONSOLE MESSAGE: line 9: Multiline
Message #85
CONSOLE MESSAGE: line 9: Multiline
Message #86
CONSOLE MESSAGE: line 9: Multiline
Message #87
CONSOLE MESSAGE: line 9: Multiline
Message #88
CONSOLE MESSAGE: line 9: Multiline
Message #89
CONSOLE MESSAGE: line 9: Multiline
Message #90
CONSOLE MESSAGE: line 9: Multiline
Message #91
CONSOLE MESSAGE: line 9: Multiline
Message #92
CONSOLE MESSAGE: line 9: Multiline
Message #93
CONSOLE MESSAGE: line 9: Multiline
Message #94
CONSOLE MESSAGE: line 9: Multiline
Message #95
CONSOLE MESSAGE: line 9: Multiline
Message #96
CONSOLE MESSAGE: line 9: Multiline
Message #97
CONSOLE MESSAGE: line 9: Multiline
Message #98
CONSOLE MESSAGE: line 9: Multiline
Message #99
CONSOLE MESSAGE: line 9: Multiline
Message #100
CONSOLE MESSAGE: line 9: Multiline
Message #101
CONSOLE MESSAGE: line 9: Multiline
Message #102
CONSOLE MESSAGE: line 9: Multiline
Message #103
CONSOLE MESSAGE: line 9: Multiline
Message #104
CONSOLE MESSAGE: line 9: Multiline
Message #105
CONSOLE MESSAGE: line 9: Multiline
Message #106
CONSOLE MESSAGE: line 9: Multiline
Message #107
CONSOLE MESSAGE: line 9: Multiline
Message #108
CONSOLE MESSAGE: line 9: Multiline
Message #109
CONSOLE MESSAGE: line 9: Multiline
Message #110
CONSOLE MESSAGE: line 9: Multiline
Message #111
CONSOLE MESSAGE: line 9: Multiline
Message #112
CONSOLE MESSAGE: line 9: Multiline
Message #113
CONSOLE MESSAGE: line 9: Multiline
Message #114
CONSOLE MESSAGE: line 9: Multiline
Message #115
CONSOLE MESSAGE: line 9: Multiline
Message #116
CONSOLE MESSAGE: line 9: Multiline
Message #117
CONSOLE MESSAGE: line 9: Multiline
Message #118
CONSOLE MESSAGE: line 9: Multiline
Message #119
CONSOLE MESSAGE: line 9: Multiline
Message #120
CONSOLE MESSAGE: line 9: Multiline
Message #121
CONSOLE MESSAGE: line 9: Multiline
Message #122
CONSOLE MESSAGE: line 9: Multiline
Message #123
CONSOLE MESSAGE: line 9: Multiline
Message #124
CONSOLE MESSAGE: line 9: Multiline
Message #125
CONSOLE MESSAGE: line 9: Multiline
Message #126
CONSOLE MESSAGE: line 9: Multiline
Message #127
CONSOLE MESSAGE: line 9: Multiline
Message #128
CONSOLE MESSAGE: line 9: Multiline
Message #129
CONSOLE MESSAGE: line 9: Multiline
Message #130
CONSOLE MESSAGE: line 9: Multiline
Message #131
CONSOLE MESSAGE: line 9: Multiline
Message #132
CONSOLE MESSAGE: line 9: Multiline
Message #133
CONSOLE MESSAGE: line 9: Multiline
Message #134
CONSOLE MESSAGE: line 9: Multiline
Message #135
CONSOLE MESSAGE: line 9: Multiline
Message #136
CONSOLE MESSAGE: line 9: Multiline
Message #137
CONSOLE MESSAGE: line 9: Multiline
Message #138
CONSOLE MESSAGE: line 9: Multiline
Message #139
CONSOLE MESSAGE: line 9: Multiline
Message #140
CONSOLE MESSAGE: line 9: Multiline
Message #141
CONSOLE MESSAGE: line 9: Multiline
Message #142
CONSOLE MESSAGE: line 9: Multiline
Message #143
CONSOLE MESSAGE: line 9: Multiline
Message #144
CONSOLE MESSAGE: line 9: Multiline
Message #145
CONSOLE MESSAGE: line 9: Multiline
Message #146
CONSOLE MESSAGE: line 9: Multiline
Message #147
CONSOLE MESSAGE: line 9: Multiline
Message #148
CONSOLE MESSAGE: line 10: hello %cworld
CONSOLE MESSAGE: line 9: Multiline
Message #0
CONSOLE MESSAGE: line 9: Multiline
Message #1
CONSOLE MESSAGE: line 9: Multiline
Message #2
CONSOLE MESSAGE: line 9: Multiline
Message #3
CONSOLE MESSAGE: line 9: Multiline
Message #4
CONSOLE MESSAGE: line 9: Multiline
Message #5
CONSOLE MESSAGE: line 9: Multiline
Message #6
CONSOLE MESSAGE: line 9: Multiline
Message #7
CONSOLE MESSAGE: line 9: Multiline
Message #8
CONSOLE MESSAGE: line 9: Multiline
Message #9
CONSOLE MESSAGE: line 9: Multiline
Message #10
CONSOLE MESSAGE: line 9: Multiline
Message #11
CONSOLE MESSAGE: line 9: Multiline
Message #12
CONSOLE MESSAGE: line 9: Multiline
Message #13
CONSOLE MESSAGE: line 9: Multiline
Message #14
CONSOLE MESSAGE: line 9: Multiline
Message #15
CONSOLE MESSAGE: line 9: Multiline
Message #16
CONSOLE MESSAGE: line 9: Multiline
Message #17
CONSOLE MESSAGE: line 9: Multiline
Message #18
CONSOLE MESSAGE: line 9: Multiline
Message #19
CONSOLE MESSAGE: line 9: Multiline
Message #20
CONSOLE MESSAGE: line 9: Multiline
Message #21
CONSOLE MESSAGE: line 9: Multiline
Message #22
CONSOLE MESSAGE: line 9: Multiline
Message #23
CONSOLE MESSAGE: line 9: Multiline
Message #24
CONSOLE MESSAGE: line 9: Multiline
Message #25
CONSOLE MESSAGE: line 9: Multiline
Message #26
CONSOLE MESSAGE: line 9: Multiline
Message #27
CONSOLE MESSAGE: line 9: Multiline
Message #28
CONSOLE MESSAGE: line 9: Multiline
Message #29
CONSOLE MESSAGE: line 9: Multiline
Message #30
CONSOLE MESSAGE: line 9: Multiline
Message #31
CONSOLE MESSAGE: line 9: Multiline
Message #32
CONSOLE MESSAGE: line 9: Multiline
Message #33
CONSOLE MESSAGE: line 9: Multiline
Message #34
CONSOLE MESSAGE: line 9: Multiline
Message #35
CONSOLE MESSAGE: line 9: Multiline
Message #36
CONSOLE MESSAGE: line 9: Multiline
Message #37
CONSOLE MESSAGE: line 9: Multiline
Message #38
CONSOLE MESSAGE: line 9: Multiline
Message #39
CONSOLE MESSAGE: line 9: Multiline
Message #40
CONSOLE MESSAGE: line 9: Multiline
Message #41
CONSOLE MESSAGE: line 9: Multiline
Message #42
CONSOLE MESSAGE: line 9: Multiline
Message #43
CONSOLE MESSAGE: line 9: Multiline
Message #44
CONSOLE MESSAGE: line 9: Multiline
Message #45
CONSOLE MESSAGE: line 9: Multiline
Message #46
CONSOLE MESSAGE: line 9: Multiline
Message #47
CONSOLE MESSAGE: line 9: Multiline
Message #48
CONSOLE MESSAGE: line 9: Multiline
Message #49
CONSOLE MESSAGE: line 9: Multiline
Message #50
CONSOLE MESSAGE: line 9: Multiline
Message #51
CONSOLE MESSAGE: line 9: Multiline
Message #52
CONSOLE MESSAGE: line 9: Multiline
Message #53
CONSOLE MESSAGE: line 9: Multiline
Message #54
CONSOLE MESSAGE: line 9: Multiline
Message #55
CONSOLE MESSAGE: line 9: Multiline
Message #56
CONSOLE MESSAGE: line 9: Multiline
Message #57
CONSOLE MESSAGE: line 9: Multiline
Message #58
CONSOLE MESSAGE: line 9: Multiline
Message #59
CONSOLE MESSAGE: line 9: Multiline
Message #60
CONSOLE MESSAGE: line 9: Multiline
Message #61
CONSOLE MESSAGE: line 9: Multiline
Message #62
CONSOLE MESSAGE: line 9: Multiline
Message #63
CONSOLE MESSAGE: line 9: Multiline
Message #64
CONSOLE MESSAGE: line 9: Multiline
Message #65
CONSOLE MESSAGE: line 9: Multiline
Message #66
CONSOLE MESSAGE: line 9: Multiline
Message #67
CONSOLE MESSAGE: line 9: Multiline
Message #68
CONSOLE MESSAGE: line 9: Multiline
Message #69
CONSOLE MESSAGE: line 9: Multiline
Message #70
CONSOLE MESSAGE: line 9: Multiline
Message #71
CONSOLE MESSAGE: line 9: Multiline
Message #72
CONSOLE MESSAGE: line 9: Multiline
Message #73
CONSOLE MESSAGE: line 9: Multiline
Message #74
CONSOLE MESSAGE: line 9: Multiline
Message #75
CONSOLE MESSAGE: line 9: Multiline
Message #76
CONSOLE MESSAGE: line 9: Multiline
Message #77
CONSOLE MESSAGE: line 9: Multiline
Message #78
CONSOLE MESSAGE: line 9: Multiline
Message #79
CONSOLE MESSAGE: line 9: Multiline
Message #80
CONSOLE MESSAGE: line 9: Multiline
Message #81
CONSOLE MESSAGE: line 9: Multiline
Message #82
CONSOLE MESSAGE: line 9: Multiline
Message #83
CONSOLE MESSAGE: line 9: Multiline
Message #84
CONSOLE MESSAGE: line 9: Multiline
Message #85
CONSOLE MESSAGE: line 9: Multiline
Message #86
CONSOLE MESSAGE: line 9: Multiline
Message #87
CONSOLE MESSAGE: line 9: Multiline
Message #88
CONSOLE MESSAGE: line 9: Multiline
Message #89
CONSOLE MESSAGE: line 9: Multiline
Message #90
CONSOLE MESSAGE: line 9: Multiline
Message #91
CONSOLE MESSAGE: line 9: Multiline
Message #92
CONSOLE MESSAGE: line 9: Multiline
Message #93
CONSOLE MESSAGE: line 9: Multiline
Message #94
CONSOLE MESSAGE: line 9: Multiline
Message #95
CONSOLE MESSAGE: line 9: Multiline
Message #96
CONSOLE MESSAGE: line 9: Multiline
Message #97
CONSOLE MESSAGE: line 9: Multiline
Message #98
CONSOLE MESSAGE: line 9: Multiline
Message #99
CONSOLE MESSAGE: line 9: Multiline
Message #100
CONSOLE MESSAGE: line 9: Multiline
Message #101
CONSOLE MESSAGE: line 9: Multiline
Message #102
CONSOLE MESSAGE: line 9: Multiline
Message #103
CONSOLE MESSAGE: line 9: Multiline
Message #104
CONSOLE MESSAGE: line 9: Multiline
Message #105
CONSOLE MESSAGE: line 9: Multiline
Message #106
CONSOLE MESSAGE: line 9: Multiline
Message #107
CONSOLE MESSAGE: line 9: Multiline
Message #108
CONSOLE MESSAGE: line 9: Multiline
Message #109
CONSOLE MESSAGE: line 9: Multiline
Message #110
CONSOLE MESSAGE: line 9: Multiline
Message #111
CONSOLE MESSAGE: line 9: Multiline
Message #112
CONSOLE MESSAGE: line 9: Multiline
Message #113
CONSOLE MESSAGE: line 9: Multiline
Message #114
CONSOLE MESSAGE: line 9: Multiline
Message #115
CONSOLE MESSAGE: line 9: Multiline
Message #116
CONSOLE MESSAGE: line 9: Multiline
Message #117
CONSOLE MESSAGE: line 9: Multiline
Message #118
CONSOLE MESSAGE: line 9: Multiline
Message #119
CONSOLE MESSAGE: line 9: Multiline
Message #120
CONSOLE MESSAGE: line 9: Multiline
Message #121
CONSOLE MESSAGE: line 9: Multiline
Message #122
CONSOLE MESSAGE: line 9: Multiline
Message #123
CONSOLE MESSAGE: line 9: Multiline
Message #124
CONSOLE MESSAGE: line 9: Multiline
Message #125
CONSOLE MESSAGE: line 9: Multiline
Message #126
CONSOLE MESSAGE: line 9: Multiline
Message #127
CONSOLE MESSAGE: line 9: Multiline
Message #128
CONSOLE MESSAGE: line 9: Multiline
Message #129
CONSOLE MESSAGE: line 9: Multiline
Message #130
CONSOLE MESSAGE: line 9: Multiline
Message #131
CONSOLE MESSAGE: line 9: Multiline
Message #132
CONSOLE MESSAGE: line 9: Multiline
Message #133
CONSOLE MESSAGE: line 9: Multiline
Message #134
CONSOLE MESSAGE: line 9: Multiline
Message #135
CONSOLE MESSAGE: line 9: Multiline
Message #136
CONSOLE MESSAGE: line 9: Multiline
Message #137
CONSOLE MESSAGE: line 9: Multiline
Message #138
CONSOLE MESSAGE: line 9: Multiline
Message #139
CONSOLE MESSAGE: line 9: Multiline
Message #140
CONSOLE MESSAGE: line 9: Multiline
Message #141
CONSOLE MESSAGE: line 9: Multiline
Message #142
CONSOLE MESSAGE: line 9: Multiline
Message #143
CONSOLE MESSAGE: line 9: Multiline
Message #144
CONSOLE MESSAGE: line 9: Multiline
Message #145
CONSOLE MESSAGE: line 9: Multiline
Message #146
CONSOLE MESSAGE: line 9: Multiline
Message #147
CONSOLE MESSAGE: line 9: Multiline
Message #148
CONSOLE MESSAGE: line 10: hello %cworld
Verifies viewport stick-to-bottom behavior.


Running: verifyViewportIsTallEnough
Viewport contains 13 messages

Running: testScrollViewportToBottom
Is at bottom: true, should stick: true

Running: testConsoleSticksToBottom
Is at bottom: true, should stick: true

Running: testSmoothScrollDoesNotStickToBottom
Is at bottom: false, should stick: false

Running: testEscShouldNotJumpToBottom
Is at bottom: false, should stick: false

Running: testTypingShouldJumpToBottom
Is at bottom: true, should stick: true

Running: testViewportMutationsShouldPreserveStickToBottom
Is at bottom: true, should stick: true
Is at bottom: false, should stick: false

Running: testMuteUpdatesWhileScrolling
New messages were muted: true
Refresh was scheduled after dirty state

Running: testShouldNotJumpToBottomWhenPromptFillsEntireViewport
Is at bottom: false, should stick: true
